app/requirements.txt
asyncio
playwright
openai
requests
aiofiles
tarsier
awslambdaric
python-dotenv
capsolver
playwright-stealth
